requirements.txt
# Core
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0

# Faker & Random Data
faker>=25.0.0
uuid

# Expression Evaluation (Safe)
simpleeval>=0.9.13

# Optional Enhancements
PyYAML>=6.0    # For YAML-based schema support
openpyxl       # For Excel export support

# Validation
pydantic>=2.0  # Optional, for schema validation (or use jsonschema)

# Dev/Debugging
python-dotenv  # For environment variable management (optional)
